We should not foster and propagate wrong concepts to programmers
the `.json()` is to parse JSON (string), and returns a JavaScript **object**.

It DOES NOT return JSON. It returns data, in the form of a JavaScript object.

There are programmers coming up, asking about a JavaScript object but calling it a json.

`.json()` is not to return JSON. `.json()` is to parse JSON and return data, in the form of a JavaScript object.  We should not foster and propagate wrong concepts to thousands and millions of programmers.