fix: update code to match tutorial
As mentioned in https://github.com/mdn/webgl-examples/issues/26 the tutorial mentions the use of `window.onload = main;` but, the code in the repo does not follow that patter. It instead calls the `main` function on line 1 of the JavaScript block.

This pull request updates the example to match the copy in the tutorial.

fix #26