Update README.md
The original text attributes the failure of running the samples from a file:/// URL to a non-secure context. The real reason is not a security issue but a CORS policy: with a file protocol scheme, the origin is 'null', and the null is identical to nothing. Therefore, the CORS policy check fails.
For Sample 8 to run, the browser can require "user gesture". I explain how to use the browser privacy settings to circumvent this requirement.

Is a Tetrahedron link dead or is it only my connectivity problem? If the link is dead, delete the 'Additional' example' subsection.